Create smokersProblem.go
The smokers problem describes a situation in which 3 smokers sit at a table. Everyone of them has tobacco, smoking paper, and matches, respectively, but do not share with each other. 
The waiter comes to the table and puts 2 smoking utensils on it. The smoker who has the complementary utensil can smoke now. He does it, stops, and then all 3 wait again for the waiter.